advanced-e2e-tests-with-react/backend/requirements.txt
anthropic[bedrock]>=0.37.1
jsonschema==4.22.0
boto3>=1.28.57
google-auth>=2.20.0,<3
python-dotenv
httpx
selenium
rich
webdriver-manager
pyyaml
